CE: Default-Initialize shadowRadius and a timestamp
The unit tests in particular assume that all values are initialzed to
some reasonable default.

The OutputTest takesEarlyOutIfLayerHasEmptyVisibleRegion was aborting on
an integer overflow inside Rect::inset() because an uninitialized
shadowRadius was being converted to -INT32_MIN.

The OutputTest gathersClientCompositionRequests() was sometimes failing because it
assumes that the timestamp is set to zero by default, and is set to a
(nonzero) value by the code under test.

With both these fixes, the CompositionEngine tests pass.

Test: atest libcompositionengine_test
Bug: None
Change-Id: I7ff56a0f24f6d5960493b18458f52766f955151d