Track focused window by using window token
A window token can be associated with one or more InputWindowHandles. If
a window is mirrored, the window and its mirror will share the same
token. Focus is tracked by the token per display and the events are
dispatched to the channel associated by this token.

Also pass window handle per display by reference where possible.

Test: atest inputflinger_tests
Bug: 151179149
Change-Id: I0c5ad25e60e85b1a8f5648c2dacb125b13e00e55