Check if input devices are opening
If input devices are opening, but not yet open, they will be sitting in
mOpeningDevices list and not in mDevices map.

Therefore, when pairing a video device, we should check both places.

This is needed because of a recent behaviour change where device can
only be in 1 place at a given time.

Bug: 165554458
Test: adb shell dumpsys input
Change-Id: Id44a2d7a6f8246dec26cb33c6d073d7f0c39e507