Fix SurfaceFlinger_test failed when display color mode changed
The screenshot will be used to verify if it contains the expected
colors in test cases. But the dataspace would depend on the color mode
of the display. And the captured screenshot could be different to the
native values and cause the test failed because it check the unexpected
color.

- Add 'useRGBColorSpace' in CaptureArgs to force capture screenshot in
  native rgb so we could use it to verify the expected color.
- getActiveConfig will return the id instead the status, we have to
  check the value if valid.

Bug: 166622214
Test: change color mode and run atest SurfaceFlinger_test
Change-Id: I2440e13f2ba0932b770158a11f2eab1903c77b30