Revert "Revert "libbinder_rs: Treat previously associated Binder as remote""
This reverts commit f6b9ad98ee61149f42b6ca0bf84068d4411a92f8.

Reason for revert: Includes fix for aidl_integration_test failure that prompted the initial revert

Original commit message:
libbinder_rs: Treat previously associated Binder as remote

A binder object may have already been associated by another module (e.g.
NDK code that created the object if it is local) and therefore already
have an NDK AIBinder_Class. In this case we still want to transact with
the object, but must treat it as remote if its class does not match the
class expected for a compatible Rust service.

Re-landing this change with the fix for aidl_integration_test failure
that prompted its reversion.

Bug: 167723746
Test: atest rustBinderTest binderRustNdkInteropTest

Change-Id: I9a848bcc4b8003a22ebe59c89a95be70ef6ee7ec