Added SKIP_SCREENSHOT flag for layers
The previous way to skip taking a layer in the screenshot was to set the
windowType to WINDOW_TYPE_DONT_SCREENSHOT. This is confusing since
skipping screenshots is not really related to window type. Instead, use
the new flag SKIP_SCREENSHOT that can be set when the layer is created
or in a transaction. This simplifies the behavior when deciding if the
layer should be screenshot.

This also fixes a bug where if the SKIP_SCREENSHOT flag was added after
creation, the layer would still be screenshot since the flag primaryDisplayOnly
was not changed.

This also makes it so the flag is inherited by the children. So if a
parent sets the flag, the children will also not be in the screenshot.

Test: ScreenCapture_test
Fixes: 168943350
Change-Id: I26d2b3d025d8daac17b303a96d5b54fc104f0d50