Limit the number of printed entries in a queue
When debugging some input bugs, it takes a while to scroll through all
of the entries in waitQueues and outboundQueues.
For example, the count can be > 100 for some ANR scenarios.

Limit the number to 50 total entries printed: 25 oldest and 25 newest.
This quantity should be sufficient to figure out what's going on with
the queues.

Note: I could not reproduce the ANR anymore, so the latest patchset was
not tested using the first test procedure.

Bug: 160561987
Test: reproduce an ANR using monkey, then look at dumpsys input
adb shell monkey 10000
adb shell dumpsys input
Test: the above procedure, but without 'reproduce ANR'

Change-Id: I01d661e0a85cc5bfbf61fa4b23689e8b27daa84b