GPU Memory: update unittest to reduce flakiness
bpf::setrlimitForTest is not a requirement but helping to reserve enough
memory on cloud Android instance for passing bpf related tests. Now the
errno has been refactored to tracked inside that API. This change just
removes the check in this unittest because failing to set the new limit
doesn't necessarily lead to the actual test case failures.

Bug: 165604786
Test: GpuMemTest
Change-Id: Ibdf8c2c5170bcfd5deb3b461eec8962c4a28954b