Add buffer rejection logic to BlastBufferQueue
If the client sets the buffer scaling mode to freeze, then we want to
reject any buffers that do not match default width and size of the
buffer.

Additionally, we add logic to wait for the first callback before
acquiring a second buffer this is to ensure we will present the first
buffer if multiple buffers are queued in succession.

Test: atest SurfaceViewBufferTests
Bug: 168504870
Change-Id: I987f8ce4047a6f7f52726cfd404fbe9d565f83a2