[SF] Set default layerStackRect to the framebuffer size
Currently we set the default layer stack rect to the size of
the screen.  In practice this default value is used only for the
first few seconds of the boot animation until DisplayManager sets
projection. However some TV devices run with a framebuffer smaller
than the display which cases the boot animatiob to be misscaled.
In this CL we change the default to the framebuffer size.
This change should have no effect for non-TV devices, since there
the framebuffer size = display size.

Bug: 169694490
Test: flash a device where ro.surface_flinger.max_graphics_heigh=1080
      and display mode is 4k, make sure the boot animation is
      displayed properly
Test: flash a device with 1080p display size and framebuffer and
      make sure the  boot animation is displayed properly
Change-Id: I91b4623bd0eab826d15abd2874929ae3758602f1