Implement initial version of tone mapping in Skia RenderEngine.
Normally Skia does not perform the same type of polynomial tone mapping
that RenderEngine is expected to do. As such, HDR playback presents far
too darkly when presented alongside SDR content on an SDR target buffer.
Furthermore, precompose HAL is still under development and is not
guaranteed to support tone mapping or be present on all devices, so Skia
RenderEngine needs to support this so that we can test internally.

Works with Youtube HDR - other formats like HLG haven't been implemented
yet and will need some testing.

This structure is also expected to support color transforms applied in
linear space in an upcoming patch.

Bug: 164223050
Test: Peru HDR
Change-Id: Ie95b58e9dad4e4f1936d49e24c426cc884d2de4d