Don't increment layer state sequence id when setting setFrameTimelineVsyncForTransaction
Layer state sequence id should only change when the visible region
can change. The id is used to set the mVisibleRegionDirty flag which
triggers geom calculations and updates to InputDispatcher.

setFrameTimelineVsyncForTransaction is called for every layer
transaction and this is worse for Blast since buffers come via
transactions as well.

Test: adb shell dumpsys SurfaceFlinger --frametimeline and check prediction state is valid
Change-Id: Ib8ba61a706d122e3e2d77730dcabf29c7480529b