Add dump of pending focus requests
Assuming we are keeping this code, let's add a dump of pending focus
requests for better debugging of focus-related issues.

Since printing a token is pretty meaningless (and not clear how to do),
we try to find the window associated with this token instead.

Sample dump:

mPendingFocusRequests:
    displayId=0, token->'Window{ff00537 u0 com.android.chrome/org.chromium.chrome.browser.firstrun.FirstRunActivity}', focusedToken->'<nullptr>'

Bug: 171130500
Test: adb shell monkey 10000 then adb shell dumpsys input | grep -i
focusreq -A 2 and see the output produced. Sample output is above

Change-Id: Id99ae4f3253a5037b352b41ee108e19c7f6f8eb2