Fix assert error on vibrator benchmarks
The benchmark Args are generated dynamically with the supported
effects/primitives, and they can be an empty list on some devices.

The benchmark.h has a range(size_t) method to retrive arguments, which
asserts on the args size, but does not expose a way to check if the args
are empty.

Fix: 171192316
Test: atest libvibratorservice_benchmarks
Change-Id: Ida67e81bdf1066a055a6cafbb3b1ef3bdbe23e41