Traverse layers in main thread.
Currently we have to traverse layers in order to determine whether a
protected buffer must be allocated, however, screenshot functions are
called in binder threads most of the time and hence we ended up
traversing layers in binder threads, which is wrong. This patch
schedules the traverse to main thread.

Bug: b/172774482
Test: Protected video is visible during rotation
Test: traverse layers are in main thread in trace
Change-Id: Iefa6fec78586af00a28f87fa5641d5fff929b80f