Support linear color transforms in Skia-RenderEngine.
Per-layer color transforms are defined by the HWC spec to apply in linear
space immediately before OETF. Support that by passing an additional
matrix into LinearEffect.

Note that this probably is not the most efficient way to do this at
runtime; when tone mapping is not required and RGB primaries are the
same, we can remove an additional matrix computation, but the number of
shaders cached on-device is cut down this way.

Bug: 164223050
Test: Hack in a color transform and compare gles vs skiagl.
Change-Id: If01995a90a4381d811c39c68cff5983b04e2971d