Use identity transform for Joystick MotionEvent
The SyntheticJoystickHandler makes key events
according to the axis values of MotionEvent.

If there is a Window which has transformed,
the axis values are going to be bigger or smaller than
SyntheticJoystickHandler's threshold and the generating
events never stop because JoystickHandler thinks that
the user is controlling Joystick continuously.

This makes Joystick's MotionEvent has identity transform
so that its aix values are not depending on the window but
display.

Test: atest inputflinger_tests:InputDispatcherTest\
        #NonPointerMotionEvent_JoystickNotTransformed
Bug: 158802274
Change-Id: Ib1c12c773a604c74807033819b8d1e1032913c98