BlastBufferQueue: Add buffer rejection
Second attempt at rejecting buffers that do not match layer size when
buffer scaling mode is set to freeze. If we dont reject the buffers,
the layer will scale the buffers incorrectly.

However we do not want to reject buffers that were rendered before the
size change. To prevent this, we gate the size change until we receive
the first buffer with the requested size.

Bug: 168504870
Test: atest SurfaceViewBufferTests libgui_test
Test: go/wm-tests, check logs for no rejected buffers

Change-Id: I3cc43b2bbf32b4fe6cd8d540967f217c0593aa57