Fix vsyncId for long frames and double stuffed
Since the vsyncId was applied immediately, and since we are doing
client-side queing in case we are double stuffed, there was a
scenario when setFrameTimelineVsyncId was applied too early
meaning that the previous frame was tagged with the id from the
next frame.

To fix that, we set the pending vsync id and then apply it
together with the buffer.

Test: Expand quick settings, ensure no missing jank data for
shared timeline jank classification

Change-Id: Id54133b84e5305d7f1f21dbfd9f0889ef3e464c9